pyproject.toml
[project]
name = "ship-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "catboost>=1.2.8",
    "faiss-cpu>=1.12.0",
    "google-genai>=1.47.0",
    "huggingface-hub==0.24.6",
    "json-repair>=0.52.4",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-postgres>=3.0.0",
    "lightgbm>=4.6.0",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.2.12",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers==2.6.1",
    "streamlit>=1.51.0",
    "tokenizers>=0.20",
    "torch>=2.2.0",
    "transformers==4.45.2",
]
